Make find_users ACTUALLY get ALL users
The previous implementation only got the first page of users.
Because find_users only returns 'values' (an array of Users) and
excludes other parts of the response, the caller cannot handle
pagination even if that was the correct place to put it.

This implementation will do the initial request to get the first
page of results, yield the results (if a block is given), then
repeatedly request odata.nextLink's value, while present, and
return a concatenated list of all users retrieved in all pages.